requirements.txt
flask
mysqlclient
PyMySQL
coverage
flask_sqlalchemy
pyyaml
nose2
